backend/requirements.txt
djangorestframework
markdown
django-markdown
django-grappelli
django-system-monitor
pillow
